refactor: vtl analysis functions
